done code refactoring
