+++ src/js/loader.js
@@ -1,0 +1,7 @@
+export const hideLoader = () => {
+    document.querySelector('#loader').style.display = 'none';
+};
+
+export const showLoader = () => {
+    document.querySelector('#loader').style.display = 'block';
+};
